+++ src/files_select.c
@@ -307,7 +307,7 @@ static int SelectOwnerMatch(char *path,struct stat *lstatptr,struct Rlist *crit)
 InitAlphaList(&leafattrib);
   
 #ifndef MINGW  // no uids on Windows
-sprintf(buffer,"%d",lstatptr->st_uid);
+sprintf(buffer,"%jd",(uintmax_t)lstatptr->st_uid);
 PrependAlphaList(&leafattrib,buffer); 
 #endif  /* MINGW */
 
@@ -571,7 +571,7 @@ static int SelectGroupMatch(struct stat *lstatptr,struct Rlist *crit)
 
 InitAlphaList(&leafattrib);
   
-sprintf(buffer,"%d",lstatptr->st_gid);
+sprintf(buffer,"%jd",(uintmax_t)lstatptr->st_gid);
 PrependAlphaList(&leafattrib,buffer); 
 
 if ((gr = getgrgid(lstatptr->st_gid)) != NULL)
